initial ZEL changes including Acadia naming

+++ src/snark/libsnark/algebra/curves/alt_bn128/alt_bn128_g1.hpp
@@ -23,8 +23,8 @@ public:
     static int64_t add_cnt;
     static int64_t dbl_cnt;
 #endif
-    static std::vector<size_t> wnaf_window_table;
-    static std::vector<size_t> fixed_base_exp_window_table;
+    static std::vector<uint64_t> wnaf_window_table;
+    static std::vector<uint64_t> fixed_base_exp_window_table;
     static alt_bn128_G1 G1_zero;
     static alt_bn128_G1 G1_one;
